% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmapi.R
\name{apply_revision}
\alias{apply_revision}
\title{Apply: Apply, or publish, revision}
\usage{
apply_revision(revision_response_object, domain_url, email, password)
}
\arguments{
\item{revision_response_object}{list. The response body from the open_revision function.}

\item{domain_url}{The Socrata domain URL, including 'https://'}

\item{email}{Socrata username, or Socrata API Key}

\item{password}{Socrata password, or Socrata API Secret}
}
\value{
This function returns the response body from the httr call to the Socrata Data Management API apply endpoint.
}
\description{
This function applies the revision and publishes the data update, once the 
data has completed its data validation and transformation steps.
}
\note{
More information on Data Management API available at https://socratapublishing.docs.apiary.io/
}
\examples{
\dontrun{
apply_revision()
}
}
