% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmapi.R
\name{create_source}
\alias{create_source}
\title{Source: Specify details about your data source}
\usage{
create_source(revision_response_object, filename, source_type = "upload",
  source_parse = "true", domain_url, email, password)
}
\arguments{
\item{revision_response_object}{The response from the API from the open_revision function.}

\item{filename}{The name of the file you are uploading.}

\item{source_type}{Default is "upload". Only uploads supported at this time.}

\item{source_parse}{'true' or 'false' for whether the source is of a data type Socrata can parse, including
csv, tsv, xls, OpenXML, GeoJSON, KML, KMZ, Shapefile. Default is 'true'.}

\item{domain_url}{The Socrata domain URL, including 'https://'}

\item{email}{Socrata username, or Socrata API Key}

\item{password}{Socrata password, or Socrata API Secret}
}
\value{
This function returns the response body from the httr call to the Socrata Data Management API's source endpoint
}
\description{
Within a revision, a Source describes pertinent details about your data source, like whether or not it can be parsed.
This function passes options to the Source endpoint of your revision, and is the second step in
running a dataset update.
}
\note{
More information on Data Management API available at https://socratapublishing.docs.apiary.io/
}
\examples{
\dontrun{
create_source()
}
}
