% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.r
\name{DBI}
\alias{DBI}
\title{DBI methods}
\description{
Implementations of functions defined in the `DBI` package. Cribbed
heavily from `bigrquery` package, and Hadley's DBI backend vignette
(https://rstats-db.github.io/DBI/articles/backend.html)
}
