% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSocrata.R
\name{export.socrata}
\alias{export.socrata}
\title{Exports CSVs from Socrata data portals}
\usage{
export.socrata(url, path = getwd(), app_token = NULL)
}
\arguments{
\item{url}{- the base URL of a domain (e.g., "data.cityofchicago.org")}

\item{app_token}{- a string; SODA API token used to query the data 
portal \url{http://dev.socrata.com/consumers/getting-started.html}}
}
\value{
a Gzipped file with the four-by-four and timestamp of when the download began in filename
}
\description{
Input the base URL of a data portal (e.g., "data.cityofchicago.org") and
will download CSVs, PDFs, Word, Excel,  and PowerPoint files contained on
the respective data portal into a single directory named after the root URL.
Downloaded CSV files are compressed to GZip format and each file timestamped 
so the download time is cataloged. The site's data.json file is downloaded 
as a canonical index of data saved from the website. Users can cross-reference 
the data.json file by matching the "four-by-four" in data.json with the first 
5 letters of downloaded files.
}
\author{
Tom Schenk Jr \email{tom.schenk@cityofchicago.org}
}
