% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSocrata.R
\name{fieldName}
\alias{fieldName}
\title{Convert Socrata human-readable column name to field name}
\usage{
fieldName(humanName)
}
\arguments{
\item{humanName}{- a Socrata human-readable column name}
}
\value{
Socrata field name
}
\description{
Convert Socrata human-readable column name,
as it might appear in the first row of data,
to field name as it might appear in the HTTP header;
that is, lower case, periods replaced with underscores#'
}
\examples{
fieldName("Number.of.Stations") # number_of_stations
}
\author{
Hugh J. Devlin, Ph. D. \email{Hugh.Devlin@cityofchicago.org}
}
