% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Return metadata about a Socrata dataset}
\source{
\url{http://stackoverflow.com/a/29782941}
}
\usage{
getMetadata(url = "", email = NULL, password = NULL)
}
\arguments{
\item{url}{- A Socrata resource URL, or a Socrata "human-friendly" URL!}

\item{optional}{email - The email to the Socrata account with read access to the dataset}

\item{optional}{password - The password associated with the email to the Socrata account}
}
\value{
a list (!) containing a number of rows & columns and a data frame of metadata
}
\description{
This function returns metadata about a dataset. Generally, such metadata can be accessed
with browser at \code{http://DOMAIN/api/views/FOUR-FOUR/rows.json} or
\code{http://DOMAIN/api/views/FOUR-FOUR/columns.json}, which is used here.
}
\examples{
\dontrun{
gM1 <- getMetadata(url = "http://data.cityofchicago.org/resource/y93d-d9e3.json")
gM3 <- getMetadata(url = "https://data.cityofchicago.org/resource/6zsd-86xi.json")
gM2 <- getMetadata(url = "https://data.cityofboston.gov/resource/awu8-dc52")
}
}
\author{
John Malc \email{cincenko@outlook.com}
}

