% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSocrata.R
\name{isFourByFour}
\alias{isFourByFour}
\title{Checks the validity of the syntax for a potential Socrata dataset Unique Identifier, also known as a 4x4.}
\usage{
isFourByFour(fourByFour)
}
\arguments{
\item{fourByFour}{- a string; character vector of length one}
}
\value{
TRUE if is valid Socrata unique identifier, FALSE otherwise
}
\description{
Will check the validity of a potential dataset unique identifier
supported by Socrata. It will provide an exception if the syntax
does not align to Socrata unique identifiers. It only checks for
the validity of the syntax, but does not check if it actually exists.
}
\author{
Tom Schenk Jr \email{tom.schenk@cityofchicago.org}
}
