% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listDatasets.R
\name{ls.socrata}
\alias{ls.socrata}
\title{List datasets available from a Socrata domain}
\usage{
ls.socrata(url = "")
}
\arguments{
\item{url}{- A Socrata URL. This simply points to the site root.}
}
\value{
an R data frame containing a listing of datasets along with
various metadata.
}
\description{
List datasets available from a Socrata domain
}
\note{
URLs such as \code{"soda.demo.socrata.com"} are not supported
}
\examples{
df <- ls.socrata(url = "http://soda.demo.socrata.com")
## df.ny <- ls.socrata("https://data.ny.gov/")
}
\author{
Peter Schmiedeskamp \email{pschmied@uw.edu}
}

