% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmapi.R
\name{open_revision}
\alias{open_revision}
\title{Revision: Open a new revision on a dataset}
\usage{
open_revision(dataset_id, action_type, domain_url, email, password)
}
\arguments{
\item{dataset_id}{character vector. The dataset's unique ID, or four by four, located at the end of a dataset's URL}

\item{action_type}{character vector. One of "update", "replace", or "delete", depending on how you want to update the dataset.}

\item{domain_url}{character vector. The Socrata domain URL, including scheme and hostname, i.e. https://data.seattle.gov}

\item{email}{character vector. ocrata username, or Socrata API Key}

\item{password}{character vector. Socrata password, or Socrata API Secret}
}
\value{
This function returns the response body from the httr call to the Socrata Data Management API revision endpoint
}
\description{
This function opens a new revision, or draft, on an existing dataset on a Socrata domain.
This is the first step in updating or replacing a dataset on Socrata with new data.
}
\note{
More information on Data Management API available at https://socratapublishing.docs.apiary.io/
}
\examples{
\dontrun{
open_revision()
}
}
