% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{posixify}
\alias{posixify}
\title{Convert Socrata calendar_date string to POSIX}
\usage{
posixify(x = "")
}
\arguments{
\item{x}{- character vector in one of two Socrata calendar_date formats}
}
\value{
a POSIX date
}
\description{
Datasets will either specify what timezone they should be interpreted in,
or you can usually assume they’re in the timezone of the publisher. See examples below too.
}
\examples{
posixify("2014-10-13T23:00:00")
posixify("09/14/2012 10:38:01 PM")
posixify("09/14/2012")
}
\author{
Hugh J. Devlin, Ph. D. \email{Hugh.Devlin@cityofchicago.org}
}
\seealso{
\url{http://dev.socrata.com/docs/datatypes/floating_timestamp.html}
}

