% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnData.R
\name{read.socrata}
\alias{read.socrata}
\title{Get a full Socrata data set as an R data frame}
\usage{
read.socrata(url = NULL, app_token = NULL, limit = 50000, domain = NULL,
  fourByFour = NULL, query = NULL)
}
\arguments{
\item{url}{- A Socrata resource URL, or a Socrata "human-friendly" URL, 
or Socrata Open Data Application Program Interface (SODA) query 
requesting a comma-separated download format (.json suffix), 
May include SoQL parameters, and it is now assumed to include SODA \code{limit} 
& \code{offset} parameters.
Either use a compelete URL or use parameters below to construct your URL.}

\item{app_token}{- a (non-required) string; SODA API token can be used to query the data 
portal \url{http://dev.socrata.com/consumers/getting-started.html}}

\item{limit}{- defaults to the max of 50000. See \url{http://dev.socrata.com/docs/paging.html}.}

\item{domain}{- A Socrata domain, e.g \url{http://data.cityofchicago.org}}

\item{fourByFour}{- a unique 4x4 identifier, e.g. "ydr8-5enu". See more \code{\link{isFourByFour}}}

\item{query}{- Based on query language called the "Socrata Query Language" ("SoQL"), see 
\url{http://dev.socrata.com/docs/queries.html}.}
}
\description{
Manages throttling and POSIX date-time conversions. We support only .json suffix.
}
\examples{
\dontrun{
df_1 <- read.socrata(url = "http://soda.demo.socrata.com/resource/4334-bgaj.csv")
df_2 <- read.socrata(domain = "http://data.cityofchicago.org/", fourByFour = "ydr8-5enu")
df_3 <- read.socrata(url = "http://data.cityofchicago.org/resource/ydr8-5enu.json")
}
}
\author{
Hugh J. Devlin, Ph. D. \email{Hugh.Devlin@cityofchicago.org}
}

