% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnData.R
\name{read.socrataGEO}
\alias{read.socrataGEO}
\title{Download GeoJSON data using geojsonio package}
\usage{
read.socrataGEO(url = "", ...)
}
\arguments{
\item{url}{- A Socrata resource URL, requiring a .geojson suffix.}

\item{...}{- other arguments from \link{geojsonio} package for geojson_read method}
}
\value{
Returns a list, which is the default option here.
}
\description{
Download GeoJSON data using geojsonio package
}
\examples{
\dontrun{
df_geo <- read.socrataGEO(url = "https://data.cityofchicago.org/resource/6zsd-86xi.geojson")
}

}

