% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmapi.R
\name{update_socrata}
\alias{update_socrata}
\title{Update a Socrata Dataset with the Data Management API}
\usage{
update_socrata(dataset_id, filepath_to_data, action_type,
  source_type = "upload", source_parse = "true", domain_url, email,
  password)
}
\arguments{
\item{dataset_id}{character vector. The dataset's unique ID, or four by four, located at the end of a dataset's URL}

\item{filepath_to_data}{character vector. Local filepath to the data.}

\item{action_type}{character vector. One of "update", "replace", or "delete", depending on how you want to update the dataset.}

\item{source_type}{Default is "upload". Only uploads supported at this time.}

\item{source_parse}{'true' or 'false' for whether the source is of a data type Socrata can parse, including
csv, tsv, xls, OpenXML, GeoJSON, KML, KMZ, Shapefile. Default is 'true'.}

\item{domain_url}{character vector. The Socrata domain URL, including scheme and hostname, i.e. https://data.seattle.gov}

\item{email}{character vector. ocrata username, or Socrata API Key}

\item{password}{character vector. Socrata password, or Socrata API Secret}
}
\value{
This function returns the response body from the httr call to the Socrata Data Management API apply endpoint.
}
\description{
In one step, complete the steps of opening a draft, uploading data, and publishing your draft. 
Set the 'action_type' argument to "update", "replace", or "delete".
}
\note{
More information on Data Management API available at https://socratapublishing.docs.apiary.io/
}
\examples{
\dontrun{
update.socrata()
}
}
