% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmapi.R
\name{upload_to_source}
\alias{upload_to_source}
\title{Upload: Upload a data source to the Socrata revision}
\usage{
upload_to_source(create_source_response_object, filepath_to_data,
  domain_url, email, password)
}
\arguments{
\item{create_source_response_object}{list. The response body from the create_source function.}

\item{filepath_to_data}{character vector. Local filepath to the data.}

\item{domain_url}{The Socrata domain URL, including 'https://'}

\item{email}{Socrata username, or Socrata API Key}

\item{password}{Socrata password, or Socrata API Secret}
}
\value{
This function returns the response body from the httr call to the Socrata Data Management API upload endpoint.
}
\description{
This function supports streaming a data file from a local filepath to the Socrata revision.
After the data is uploaded to the Socrata revision, it undergoes a validation step.
}
\note{
More information on Data Management API available at https://socratapublishing.docs.apiary.io/
}
\examples{
\dontrun{
upload_to_source()
}
}
