% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeData.R
\name{write.socrata}
\alias{write.socrata}
\title{Write data to a Socrata resource}
\usage{
write.socrata(dataframe, dataset_json_endpoint, update_mode, email, password,
  app_token = NULL)
}
\arguments{
\item{dataframe}{- dataframe to upload to Socrata}

\item{dataset_json_endpoint}{- Socrata Open Data Application Program Interface (SODA) endpoint (JSON only for now)}

\item{update_mode}{- "UPSERT" or "REPLACE"; consult http://dev.socrata.com/publishers/getting-started.html}

\item{email}{- email associated with Socrata account (will need write access to dataset)}

\item{password}{- password associated with Socrata account (will need write access to dataset)}

\item{app_token}{- optional app_token associated with Socrata account}
}
\value{
httr a response object
}
\description{
Write data to a Socrata resource
}

