% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publisher.r
\name{write_socrata}
\alias{write_socrata}
\title{Publish a dataframe to Socrata}
\usage{
write_socrata(dataframe, fourfour = NULL, name = NULL, description = NULL,
  domain, username, password)
}
\arguments{
\item{dataframe}{A dataframe to publish}

\item{name}{Dataset name; ignored if overwriting an existing dataset}

\item{description}{Dataset description; ignored if overwriting existing dataset}

\item{domain}{Socrata domain}

\item{username}{Username (email address) to be used for authentication}

\item{password}{Password to be used for authentication}
}
\value{
URL string pointing to newly published dataset
}
\description{
Publish a dataframe to Socrata
}
\author{
Peter Schmiedeskamp \email{peter.schmiedeskamp@socrata.com}
}
