% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSocrata.R
\name{export.socrata}
\alias{export.socrata}
\title{Exports CSVs from Socrata data portals}
\usage{
export.socrata(url, app_token = NULL)
}
\arguments{
\item{url}{- the base URL of a domain (e.g., "data.cityofchicago.org")}

\item{app_token}{- a string; SODA API token used to query the data 
portal \url{http://dev.socrata.com/consumers/getting-started.html}}
}
\value{
a Gzipped file with the four-by-four and timestamp of when the download began in filename
}
\description{
Input the URL of a data portal (e.g., "data.cityofchicago.org") and
will download all CSV files (no other files supported) and saved in
a single directory named after the root URL (e.g., "data.cityofchicago.org/").
Downloaded files are compressed to GZip format and timestamped so the download
time is saved. No data is saved within the R workspace.
}
\author{
Tom Schenk Jr \email{tom.schenk@cityofchicago.org}
}
