% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Return metadata about a Socrata dataset}
\source{
\url{http://stackoverflow.com/a/29782941}
}
\usage{
getMetadata(url = "")
}
\arguments{
\item{url}{- A Socrata resource URL, or a Socrata "human-friendly" URL}
}
\value{
a list containing a number of rows and a data frame of metadata
}
\description{
This can be either in format
'http://data.cityofchicago.org/api/views/xzkq-xp2w/rows.json' or
'http://data.cityofchicago.org/api/views/xzkq-xp2w/columns.json', which this function also uses.
}
\examples{
getMetadata(url = "http://data.cityofchicago.org/resource/y93d-d9e3.csv")
getMetadata(url = "https://data.cityofchicago.org/resource/6zsd-86xi.json")
}

