% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Return metadata about a Socrata dataset}
\source{
\url{http://stackoverflow.com/a/29782941}
}
\usage{
getMetadata(url = "")
}
\arguments{
\item{url}{- A Socrata resource URL, or a Socrata "human-friendly" URL,
or Socrata Open Data Application Program Interface (SODA) query
requesting a comma-separated download format (.csv suffix),
May include SoQL parameters, and it is now assumed to include SODA \code{limit}
& \code{offset} parameters.
Either use a compelete URL or use parameters below to construct your URL.}
}
\value{
a list (!) containing a number of rows and a data frame of metadata
}
\description{
This function returns metadata about a dataset. Generally, such metadata can be accessed
with browser at \code{http://DOMAIN/api/views/FOUR-FOUR/rows.json} or
\code{http://DOMAIN/api/views/FOUR-FOUR/columns.json} (this is used here).
}
\examples{
\dontrun{
asd <- getMetadata(url = "http://data.cityofchicago.org/resource/y93d-d9e3.csv")
getMetadata(url = "https://data.cityofchicago.org/resource/6zsd-86xi.json")
wqfef <- getMetadata(url = "https://data.cityofboston.gov/resource/awu8-dc52")
}
}

