% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSocrata.R
\name{ls.socrata}
\alias{ls.socrata}
\title{List datasets available from a Socrata domain}
\usage{
ls.socrata(url)
}
\arguments{
\item{url}{- A Socrata URL. This simply points to the site root.}
}
\value{
an R data frame containing a listing of datasets along with
various metadata.
}
\description{
List datasets available from a Socrata domain
}
\examples{
# Download list of data sets
df <- ls.socrata("https://soda.demo.socrata.com")
# Check schema definition for metadata
attributes(df)
}
\author{
Peter Schmiedeskamp \email{pschmied@uw.edu}
}
