% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSocrata.R
\name{posixify}
\alias{posixify}
\title{Convert Socrata calendar_date string to POSIX}
\usage{
posixify(x)
}
\arguments{
\item{x}{- character vector in one of two Socrata calendar_date formats}
}
\value{
a POSIX date
}
\description{
Convert Socrata calendar_date string to POSIX
}
\author{
Hugh J. Devlin, Ph. D. \email{Hugh.Devlin@cityofchicago.org}
}
