\name{posixify}
\alias{posixify}
\title{Convert Socrata calendar_date string to POSIX}
\usage{
  posixify(x)
}
\arguments{
  \item{x}{a string in Socrata calendar_date format}
}
\value{
  a POSIX date
}
\description{
  Convert Socrata calendar_date string to POSIX
}
\author{
  Hugh J. Devlin, Ph. D.
  \email{Hugh.Devlin@cityofchicago.org}
}

