% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RSocrata.R
\name{read.socrata}
\alias{read.socrata}
\title{Get a full Socrata data set as an R data frame}
\usage{
read.socrata(url, app_token = NULL)
}
\arguments{
\item{url}{- A Socrata resource URL,
or a Socrata "human-friendly" URL,
or Socrata Open Data Application Program Interface (SODA) query
requesting a comma-separated download format (.csv suffix),
May include SoQL parameters,
but is assumed to not include a SODA offset parameter}

\item{app_token}{- a string; SODA API token used to query the data
portal \url{http://dev.socrata.com/consumers/getting-started.html}}
}
\value{
an R data frame with POSIX dates
}
\description{
Manages throttling and POSIX date-time conversions
}
\examples{
df <- read.socrata("http://soda.demo.socrata.com/resource/4334-bgaj.csv")
}
\author{
Hugh J. Devlin, Ph. D. \email{Hugh.Devlin@cityofchicago.org}
}

