% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/returnData.R
\name{read.socrata}
\alias{read.socrata}
\title{Get a full Socrata data set as an R data frame}
\usage{
read.socrata(url = NULL, app_token = NULL, limit = 50000, domain = NULL,
  fourByFour = NULL, query = NULL, offset = 0, output = "csv")
}
\arguments{
\item{url}{- A Socrata resource URL, or a Socrata "human-friendly" URL,
or Socrata Open Data Application Program Interface (SODA) query
requesting a comma-separated download format (.csv suffix),
May include SoQL parameters, and it is now assumed to include SODA \code{limit}
& \code{offset} parameters.
Either use a compelete URL or use parameters below to construct your URL.}

\item{app_token}{- a (non-required) string; SODA API token can be used to query the data
portal \url{http://dev.socrata.com/consumers/getting-started.html}}

\item{limit}{- defaults to the max of 50000. See \url{http://dev.socrata.com/docs/paging.html}.}

\item{domain}{- A Socrata domain, e.g \url{http://data.cityofchicago.org}}

\item{fourByFour}{- a unique 4x4 identifier, e.g. "ydr8-5enu". See more \code{\link{isFourByFour}}}

\item{query}{- Based on query language called the "Socrata Query Language" ("SoQL"), see
\url{http://dev.socrata.com/docs/queries.html}.}

\item{offset}{- defaults to the max of 0. See \url{http://dev.socrata.com/docs/paging.html}.}

\item{output}{- defaults to csv; one of \code{"csv" or "json"}.}
}
\value{
a data frame with POSIX dates if csv or json.
}
\description{
Manages throttling and POSIX date-time conversions.
}
\section{TODO}{
 \url{https://github.com/Chicago/RSocrata/issues/14}
}

\section{Issue}{
 If you get something like \code{Error in rbind(deparse.level, ...) :
numbers of columns of arguments do not match} when using "json" output, this is a known bug
\url{https://github.com/Chicago/RSocrata/issues/19}! Use instead csv output for time being.
}
\examples{
df_csv <- read.socrata(url = "http://soda.demo.socrata.com/resource/4334-bgaj.csv")
df_manual <- read.socrata(domain = "http://data.cityofchicago.org/", fourByFour = "ydr8-5enu")
df_manual2 <- read.socrata(domain = "http://data.cityofchicago.org/", fourByFour = "ydr8-5enu")
df_manual3 <- read.socrata(domain = "http://data.cityofchicago.org/", fourByFour = "ydr8-5enu",
output = "csv")
}
\author{
Hugh J. Devlin, Ph. D. \email{Hugh.Devlin@cityofchicago.org}
}

