% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/returnData.R
\name{read.socrataGEO}
\alias{read.socrataGEO}
\title{Download GeoJSON data using geojsonio package}
\usage{
read.socrataGEO(url = NULL, limit = 50000, offset = 0, method = "local",
  what = "sp", parse = FALSE, ...)
}
\arguments{
\item{what}{- What to return format is choosen. One of list or \code{\link{sp}}. Default: list.}

\item{parse}{- Parse geojson to data.frame like structures if possible or not. Default: FALSE (not)
\link{geojsonio}}
}
\value{
Returns a \code{\link{sp}}, which is the default option here.
}
\description{
Download GeoJSON data using geojsonio package
}
\examples{
df_geo <- read.socrataGEO(url = "https://data.cityofchicago.org/resource/6zsd-86xi.geojson")
}

