% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/returnData.R
\name{read.socrataGEO}
\alias{read.socrataGEO}
\title{Download GeoJSON data using geojsonio package}
\usage{
read.socrataGEO(url = "", method = "local", what = "sp", parse = FALSE,
  ...)
}
\arguments{
\item{url}{- A Socrata resource URL, or a Socrata "human-friendly" URL,
requesting a .geojson suffix.}

\item{method}{- \link{geojsonio} One of "web" or "local" (default). Matches on partial strings.}

\item{what}{- \link{geojsonio} What to return format is choosen. One of list (default) or \code{sp}.}

\item{parse}{- \link{geojsonio} Parse geojson to data.frame like structures if possible or not. Default: FALSE (~not)}

\item{...}{- other arguments from \link{geojsonio} package for geojson_read method}
}
\value{
Returns a \code{sp} object, which is the default option here.
}
\description{
Download GeoJSON data using geojsonio package
}
\examples{
\dontrun{
df_geo <- read.socrataGEO(url = "https://data.cityofchicago.org/resource/6zsd-86xi.geojson")
}
}

