% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/returnData.R
\name{read.socrataGEO}
\alias{read.socrataGEO}
\title{Download GeoJSON data using geojsonio package}
\usage{
read.socrataGEO(url = NULL, limit = 50000, offset = 0, method = "local",
  what = "sp", parse = FALSE, ...)
}
\arguments{
\item{url}{- A Socrata resource URL, or a Socrata "human-friendly" URL,
or Socrata Open Data Application Program Interface (SODA) query
requesting a comma-separated download format (.csv suffix),
May include SoQL parameters, and it is now assumed to include SODA \code{limit}
& \code{offset} parameters.
Either use a compelete URL or use parameters below to construct your URL.}

\item{limit}{- defaults to the max of 50000. See \url{http://dev.socrata.com/docs/paging.html}.}

\item{offset}{- defaults to 0. See \url{http://dev.socrata.com/docs/paging.html}.}

\item{method}{- \link{geojsonio} One of "web" or "local" (default). Matches on partial strings.}

\item{what}{- \link{geojsonio} What to return format is choosen. One of list (default) or \code{\link{sp}}.}

\item{parse}{- \link{geojsonio} Parse geojson to data.frame like structures if possible or not. Default: FALSE (~not)}

\item{...}{- other arguments from \link{geojsonio} package for geojson_read method}
}
\value{
Returns a \code{\link{sp}} object, which is the default option here.
}
\description{
Download GeoJSON data using geojsonio package
}
\examples{
\dontrun{
df_geo <- read.socrataGEO(url = "https://data.cityofchicago.org/resource/6zsd-86xi.geojson")
}
}

