% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSocrata.R
\name{validateUrl}
\alias{validateUrl}
\title{Convert, if necessary, URL to valid REST API URL supported by Socrata.}
\usage{
validateUrl(url, app_token)
}
\arguments{
\item{url}{- a string; character vector of length one}

\item{app_token}{- a string; SODA API token used to query the data 
portal \url{http://dev.socrata.com/consumers/getting-started.html}}
}
\value{
a - valid Url
}
\description{
Will convert a human-readable URL to a valid REST API call
supported by Socrata. It will accept a valid API URL if provided
by users and will also convert a human-readable URL to a valid API
URL. Will accept queries with optional API token as a separate
argument or will also accept API token in the URL query. Will
resolve conflicting API token by deferring to original URL.
}
\author{
Tom Schenk Jr \email{tom.schenk@cityofchicago.org}
}

