% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeData.R
\name{write.socrata}
\alias{write.socrata}
\title{Write to a Socrata dataset (full replace or upsert)}
\usage{
write.socrata(dataframe, dataset_json_endpoint, update_mode, email, password,
  app_token = NULL)
}
\arguments{
\item{dataframe}{- dataframe to upload to Socrata}

\item{dataset_json_endpoint}{- Socrata Open Data Application Program Interface (SODA) endpoint (JSON only for now)}

\item{update_mode}{- "UPSERT" or "REPLACE"; consult http://dev.socrata.com/publishers/getting-started.html}

\item{email}{- The email to the Socrata account with read access to the dataset}

\item{password}{- The password associated with the email to the Socrata account}

\item{app_token}{- a (non-required) string; SODA API token can be used to query the data
portal \url{http://dev.socrata.com/consumers/getting-started.html}}
}
\description{
Method for updating Socrata datasets
}
\author{
Mark Silverberg \email{mark.silverberg@socrata.com}
}

