% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSocrata.R
\name{write.socrata}
\alias{write.socrata}
\title{Write to a Socrata dataset (full replace or upsert)}
\usage{
write.socrata(dataframe, dataset_json_endpoint, update_mode, email,
  password, app_token = NULL)
}
\arguments{
\item{dataframe}{- dataframe to upload to Socrata}

\item{dataset_json_endpoint}{- Socrata Open Data Application Program Interface (SODA) endpoint (JSON only for now)}

\item{update_mode}{- "UPSERT" or "REPLACE"; consult http://dev.socrata.com/publishers/getting-started.html}

\item{email}{- The email to the Socrata account with read access to the dataset}

\item{password}{- The password associated with the email to the Socrata account}

\item{app_token}{- a (non-required) string; SODA API token can be used to query the data 
portal \url{http://dev.socrata.com/consumers/getting-started.html}}
}
\description{
Method for updating Socrata datasets with the Socrata Open Data API (SODA). 

Note: SODA differs in functionality from the Socrata Data Management API.
The Data Management API stages and transforms data in drafts on Socrata during the data ingress cycle. With any data updates to Socrata datasets that utilize
one or more on-platform data transformations, use the Data Management API to continue to apply those transformations to any new data.

If you do not need or use the on-platform transform functionality for a given dataset, the SODA API and \code{write.socrata()}
will continue to perform as expected. For more on the Socrata Data Management API, see the latter API's documentation \url{https://socratapublishing.docs.apiary.io/#}
}
\examples{
# Store user email and password
socrataEmail <- Sys.getenv("SOCRATA_EMAIL", "mark.silverberg+soda.demo@socrata.com")
socrataPassword <- Sys.getenv("SOCRATA_PASSWORD", "7vFDsGFDUG")

datasetToAddToUrl <- "https://soda.demo.socrata.com/resource/xh6g-yugi.json" # dataset

# Generate some data
x <- sample(-1000:1000, 1)
y <- sample(-1000:1000, 1)
df_in <- data.frame(x,y)

# Upload to Socrata
write.socrata(df_in,datasetToAddToUrl,"UPSERT",socrataEmail,socrataPassword)
}
\author{
Mark Silverberg \email{mark.silverberg@socrata.com}
}
